Added NodeType classes
